dispatcher/requirements.txt
selenium
python-dotenv
undetected-chromedriver
setuptools
google-cloud-pubsub
Flask
requests
gunicorn